refactor: change folder name from kebab-case to camel-case
