+++ @udir-design/react/src/index.ts
@@ -8,9 +8,9 @@ export * from './checkbox/Checkbox';
 export * from './chip/Chip';
 export * from './divider/Divider';
 export * from './dropdown/Dropdown';
-export * from './error-summary/ErrorSummary';
+export * from './errorSummary/ErrorSummary';
 export * from './fieldset/Fieldset';
-export * from './help-text/HelpText';
+export * from './helpText/HelpText';
 export * from './link/Link';
 export * from './list/List';
 export * from './loaders';
@@ -20,13 +20,13 @@ export * from './popover/Popover';
 export * from './radio/Radio';
 export * from './search/Search';
 export * from './select/Select';
-export * from './skip-link/SkipLink';
+export * from './skipLink/SkipLink';
 export * from './switch/Switch';
 export * from './table/Table';
 export * from './tabs/Tabs';
 export * from './tag/Tag';
 export * from './textarea/Textarea';
 export * from './textfield/Textfield';
-export * from './toggle-group/ToggleGroup';
+export * from './toggleGroup/ToggleGroup';
 export * from './tooltip/Tooltip';
 export * from './typography';
