+++ @udir-design/react/src/toggleGroup/ToggleGroup.stories.tsx
@@ -2,7 +2,6 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { ToggleGroup } from './ToggleGroup';
 
 const meta: Meta<typeof ToggleGroup> = {
-  title: 'ToggleGroup',
   component: ToggleGroup,
 };
